+++ FlowerForestAPI/FlowerForestAPI/Controllers/PlantController.cs
@@ -1,6 +1,10 @@
-﻿using FlowerForestAPI.DbContexts;
+﻿using FlowerForestAPI.AuthorizeUserServices;
+using FlowerForestAPI.DbContexts;
+using FlowerForestAPI.DTOs;
 using FlowerForestAPI.Models;
+using FlowerForestAPI.Repositories.CatalogueRepositories;
 using FlowerForestAPI.Repositories.PlantRepositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System;
@@ -12,16 +16,23 @@ namespace FlowerForestAPI.Controllers
 {
     [Route("api/{controller}")]
     [ApiController]
+    [Authorize]
     public class PlantController : ControllerBase
     {
         private readonly IPlantRepository plantRepository;
 
-        public PlantController(IPlantRepository plantRepository)
+        private readonly IAuthorizeUserService authorizeUserService;
+
+        public PlantController(IPlantRepository plantRepository,
+            IAuthorizeUserService authorizeUserService)
         {
             this.plantRepository = plantRepository;
+
+            this.authorizeUserService = authorizeUserService;
         }
 
         [HttpGet]
+        [Authorize] // To be authorized by roles
         public IActionResult GetPlants()
         {
             return Ok(plantRepository.GetPlants());
@@ -29,27 +40,73 @@ namespace FlowerForestAPI.Controllers
 
         [Route("{id}")]
         [HttpGet]
-        public IActionResult GetPlantById(Guid id)
+        public async Task<IActionResult> GetPlantById(Guid id)
         {
-            return Ok(plantRepository.GetPlantById(id));
+            var response = plantRepository.GetPlantById(id);
+            var plant = (PlantDTO)(response.Data);
+
+            var catalogueIsPublicAuthorizationResult = await authorizeUserService.AuthorizePublicCatalogueById(User, plant.CatalogueId);
+
+            if (catalogueIsPublicAuthorizationResult.Succeeded)
+                return Ok(response);
+
+            var authorizationResult = await authorizeUserService.AuthorizeCatalogueUserId(User, plant.CatalogueId);
+
+            if (authorizationResult.Succeeded)
+                return Ok(response);
+
+            return new ForbidResult();
+        }
+
+        [Route("Catalogue/{id}")]
+        [HttpGet]
+        public async Task<IActionResult> GetPlantsByCatalogueId([FromRoute] Guid id)
+        {
+            var catalogueIsPublicAuthorizationResult = await authorizeUserService.AuthorizePublicCatalogueById(User, id);
+
+            if (catalogueIsPublicAuthorizationResult.Succeeded)
+                return Ok(plantRepository.GetPlantsByCatalogueId(id));
+
+
+            var authorizationResult = await authorizeUserService.AuthorizeCatalogueUserId(User, id);
+
+            if (authorizationResult.Succeeded)
+                return Ok(plantRepository.GetPlantsByCatalogueId(id));
+
+            return new ForbidResult();
         }
 
         [HttpPost]
-        public IActionResult AddPlant(Plant plant)
-        {            
-            return Ok(plantRepository.AddPlant(plant));
+        public async Task<IActionResult> AddPlant([FromBody] Plant plant)
+        {
+            var authorizationResult = await authorizeUserService.AuthorizeCatalogueUserId(User, plant.CatalogueId);
+
+            if (authorizationResult.Succeeded)
+                return Ok(plantRepository.AddPlant(plant));
+
+            return new ForbidResult();
         }
 
         [HttpPut]
-        public IActionResult UpdatePlant(Plant plant)
+        public async Task<IActionResult> UpdatePlant([FromBody] Plant plant)
         {
-            return Ok(plantRepository.UpdatePlant(plant));
+            var authorizationResult = await authorizeUserService.AuthorizeCatalogueUserId(User, plant.CatalogueId);
+
+            if (authorizationResult.Succeeded)
+                return Ok(plantRepository.UpdatePlant(plant));
+
+            return new ForbidResult();
         }
 
         [HttpDelete]
-        public IActionResult DeletePlant(Plant plant)
+        public async Task<IActionResult> DeletePlant([FromBody] Plant plant)
         {
-            return Ok(plantRepository.DeletePlant(plant));
+            var authorizationResult = await authorizeUserService.AuthorizeCatalogueUserId(User, plant.CatalogueId);
+
+            if (authorizationResult.Succeeded)
+                return Ok(plantRepository.DeletePlant(plant));
+
+            return new ForbidResult();
         }
     }
 }
